==> notebooks/birdsong/04.0-birsong-statistics.ipynb <==
# View statistics for birdsong dataframes
1. load datasets
...

from parallelspaper.config.paths import DATA_DIR, FIGURE_DIR
from parallelspaper.birdsong_datasets import BCOL_DICT
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from parallelspaper.utils import save_fig
import seaborn as sns
%matplotlib inline

CAVICATH_stats_df = pd.read_pickle(DATA_DIR / 'stats_df/CAVICATH_stats_df.pickle')
Starling_stats_df = pd.read_pickle(DATA_DIR / 'stats_df/starling_stats_df.pickle')
BF_stats_df = pd.read_pickle(DATA_DIR / 'stats_df/BF_stats_df.pickle')

stats_df = pd.concat([CAVICATH_stats_df, Starling_stats_df, BF_stats_df])

letters = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z']

bw = 0.25
yoff = -0.20

fig, axs = plt.subplots(ncols=4, figsize=(20, 3))
kwk = {"lw": 6, "bw": bw}
for ix, (idx, row) in enumerate(stats_df.sort_values(by="species").iterrows()):
    ax = axs.flatten()[ix]
    ax.annotate(letters[ix], xy=(-0.05, 1.05), xycoords="axes fraction", size=20, fontweight='bold', fontfamily='Arial')
    ax.hist(
        np.array(row.isi)[np.array(row.isi) < 60],
        density=True,
        bins=np.arange(0, 60, 2),
        color=BCOL_DICT[row.species],
    )
    ax.set_xlim([0, 60])
    # ax.set_title(row.species)
    ax.set_yscale("log")
    ax.set_xlabel("Intersyllable interval (sec.)", fontsize=18)

    ax.tick_params(axis="both", labelsize=14, pad=15)
    for axis in ["top", "bottom", "left", "right"]:
        ax.spines[axis].set_linewidth(3)
        ax.spines[axis].set_color("k")
    ax.grid(False)
    ax.tick_params(which="both", direction="in", labelsize=14, pad=10)
    ax.tick_params(which="major", length=10, width=3)
    ax.tick_params(which="minor", length=5, width=2)


axs[0].set_ylabel("Prob. Density", labelpad=5, fontsize=18)
axs[0].yaxis.set_label_coords(yoff, 0.5)

save_fig(FIGURE_DIR / "syllable_isi")
