==> notebooks/birdsong/06.0-birdsong-results-df.ipynb <==
# create a results dataframe and table to export to latex

import pandas as pd
from parallelspaper.config.paths import DATA_DIR
import numpy as np
from parallelspaper import model_fitting as mf

### load data

MI_DF = pd.read_pickle((DATA_DIR / 'MI_DF/birdsong/birdsong_MI_DF_fitted.pickle'))

subset_MI_DF = MI_DF[MI_DF.type == 'day']

subset_MI_DF

### Calculate R2

R2 = subset_MI_DF[['R2_exp', 'R2_concat', 'R2_power']]
R2.columns = ['exp', 'composite', 'power-law']
R2.index = subset_MI_DF.species.values
R2 = R2.T
R2

#### AICc

AICcs = subset_MI_DF[['AICc_exp', 'AICc_concat', 'AICc_power']]
AICcs.index = subset_MI_DF.species.values
AICcs.columns = ['exp', 'composite', 'power-law']
AICcs.index = subset_MI_DF.species.values
AICcs = AICcs.T
AICcs

#### $\Delta$AIC

delta_AICcs = AICcs.T - np.repeat(np.min(AICcs.T.values, axis=1),3).reshape(4,3)
delta_AICcs = delta_AICcs.T
delta_AICcs

#### Relative likelihood

relative_likelihoods = mf.relative_likelihood(delta_AICcs)
relative_likelihoods

#### Relative probability of models

prob_models = mf.Prob_model_Given_data_and_models(relative_likelihoods)
prob_models

#### Join tables

AICcs['superlabel'] = 'AICc'
relative_likelihoods['superlabel'] = 'Relative likelihood'
R2['superlabel'] = '$r^2$'
prob_models['superlabel'] = 'Relative probability'

results_table = pd.concat([
    AICcs,
    R2,
    relative_likelihoods,
    prob_models
]).round(3).replace(0, '<0.001').replace(1, '>0.999')
results_table[''] = results_table.index
results_table.set_index(['superlabel', ''], inplace=True)
results_table

#### print in latex format

results_string = results_table.to_latex(bold_rows=True, escape=False)\
      .replace('>', '$>$')\
      .replace('<', '$<$')\
      .replace('superlabel', '')\
     .replace('\n\\textbf', '\n\midrule\n\\textbf')
print(results_string)

### Curvature minimum and maximum

CAVICATH_stats_df = pd.read_pickle(DATA_DIR / 'stats_df/CAVICATH_stats_df.pickle')
Starling_stats_df = pd.read_pickle(DATA_DIR / 'stats_df/starling_stats_df.pickle')
BF_stats_df = pd.read_pickle(DATA_DIR / 'stats_df/BF_stats_df.pickle')

stats_df = pd.concat([CAVICATH_stats_df, Starling_stats_df, BF_stats_df])

stats_df

curvature_dist = np.logspace(0,np.log10(100), base=10, num=1000)

for idx, row in subset_MI_DF.iterrows():
    birdrow = stats_df.species.values == row.species
    median_syllable_len = np.median(stats_df[birdrow].syllable_duration_s.values[0])
    median_isi =  np.median(stats_df[birdrow].isi.values[0])
    curvature_len = curvature_dist[int(row.min_peak)]
    curv_max = curvature_dist[np.argmax(row.curvature)]
    print(row.species)
    #print(curvature_len,median_syllable_len)
    print('\tmin curv phones:', round(curvature_len,3))
    print('\tmin curv seconds:', round(curvature_len*(median_syllable_len+median_isi),3))
    print('\tmax curv phones:', round(curv_max,3))
    print('\tmax curv seconds:', round(curv_max*(median_syllable_len+median_isi),3))
    print('\tsylls per bout:', np.median(stats_df[birdrow].recording_duration_syllable.values[0]))
