==> notebooks/birdsong/19.0-day-length-bouts.ipynb <==
# get bouts per day for each bird
1. load datasets
...

import pandas as pd
import numpy as np
from parallelspaper.config.paths import DATA_DIR
from parallelspaper.birdsong_datasets import MI_seqs, compress_seq

from parallelspaper.quickplots import plot_model_fits
import matplotlib.pyplot as plt
%matplotlib inline

### Load data

starling_seq_df = pd.read_pickle(DATA_DIR / 'song_seq_df/starling.pickle')
CAVI_CATH_seq_df = pd.read_pickle(DATA_DIR / 'song_seq_df/CAVI_CATH.pickle')
BF_seq_df = pd.read_pickle(DATA_DIR / 'song_seq_df/BF.pickle')

seq_dfs = pd.concat([starling_seq_df, CAVI_CATH_seq_df, BF_seq_df])

# sequence lengths
seq_dfs['sequence_lens'] = [len(i) for i in seq_dfs.syllables]
# recording number as integer
seq_dfs['rec_num'] = seq_dfs.rec_num.values.astype('int32')
# sort sequences
seq_dfs = seq_dfs.sort_values(by=['species', 'bird', 'rec_num'])
# get rid of unID'd birds (CAVI, CATH)
seq_dfs = seq_dfs[seq_dfs.bird != '?']

seq_dfs[:3]

# for each bird, for each day, count the number of bouts in that day

from tqdm.autonotebook import tqdm

day_bouts = {}
for species in (np.unique(seq_dfs.species)):
    day_bouts[species] = []
    species_df = seq_dfs[seq_dfs.species == species]
    for bird in tqdm(np.unique(species_df.bird.values)):
        bird_df = species_df[species_df.bird == bird]
        for day in np.unique(bird_df.day.values):
            day_df = bird_df[bird_df.day==day]
            day_bouts[species].append(len(day_df)) 

fig, axs = plt.subplots(ncols=4, figsize=(16,4))
for si, species in enumerate(np.unique(seq_dfs.species)):
    axs[si].hist(day_bouts[species])

np.median(day_bouts['BF'])

np.median(day_bouts['CAVI'])

np.median(day_bouts['CATH'])

np.median(day_bouts['Starling'])

np.median(np.concatenate([day_bouts[i] for i in day_bouts.keys()]))
